backend/rag_pipeline/requirements.txt
# rag_pipeline/requirements.txt
cocoindex[embeddings]
gitingest
python-dotenv
transformers
torch
sentence-transformers
neo4j
pinecone
sqlalchemy
psycopg2-binary